SPECIFICATION.md: binary signature must not be quoted.
If the sf field specifies the signature to be in binary format, quotes are ambiguous. They might be quoting the binary because there are spaces (0x20), or they might be the actual binary data itself.

To avoid this confusion, it is specified that in the case of binary data, quotation will be absent.

Signed-off-by: named-dodo <130215588+named-dodo@users.noreply.github.com>